Updated the dependencies Replaced npm with yarn Added a new contract
