+++ hardhat.config.js
@@ -6,5 +6,5 @@ require("@nomiclabs/hardhat-waffle");
 require("./tasks/faucet");
 
 module.exports = {
-  solidity: "0.7.3"
+  solidity: "0.8.3"
 };
