+++ scripts/_utils.js
@@ -1,0 +1,27 @@
+const fs = require("fs");
+
+
+// We also save the contract's artifacts and address in the frontend directory
+function saveFrontendFiles(token) {
+  const contractsDir = __dirname + "/../frontend/src/contracts";
+  if (!fs.existsSync(contractsDir)) {
+    fs.mkdirSync(contractsDir);
+  }
+
+  fs.writeFileSync(
+    contractsDir + "/contract-address.json",
+    JSON.stringify({ Token: token.address }, undefined, 2)
+  );
+
+  const TokenArtifact = artifacts.readArtifactSync("Token");
+
+  fs.writeFileSync(
+    contractsDir + "/Token.json",
+    JSON.stringify(TokenArtifact, null, 2)
+  );
+}
+
+
+module.exports = {
+  saveFrontendFiles,
+};
